+++ crates/objc2/src/encode.rs
@@ -68,7 +68,8 @@ use core::num::{
 use core::ptr::NonNull;
 use core::sync::atomic;
 
-use crate::runtime::Bool;
+// Intentionally not `#[doc(hidden)]`
+pub mod __unstable;
 
 #[doc(inline)]
 pub use objc2_encode::{Encoding, EncodingBox, ParseError};
@@ -216,8 +217,8 @@ pub unsafe trait RefEncode {
 /// See [the nomicon][nomicon-nullable] for details on which types uphold this
 /// promise.
 ///
-/// This is used to work around, which would normally prevent from
-/// implementing [`Encode`]/[`RefEncode`] for `Option<CustomType>`.
+/// This is used to work around the orphan rule, which would normally prevent
+/// you from implementing [`Encode`]/[`RefEncode`] for `Option<CustomType>`.
 ///
 /// [nomicon-nullable]: https://doc.rust-lang.org/nightly/nomicon/ffi.html#the-nullable-pointer-optimization
 ///
@@ -299,37 +300,6 @@ encode_impls!(
 );
 
 // TODO: Structs in core::arch?
-
-/// To allow usage as the return type of generic functions.
-///
-/// You should not rely on this encoding to exist for any other purpose (since
-/// `()` is not FFI-safe)!
-// TODO: Figure out a way to remove this - maybe with a `EncodeReturn` trait?
-unsafe impl Encode for () {
-    const ENCODING: Encoding = Encoding::Void;
-}
-
-// UI tests of this is too brittle.
-#[cfg(doctest)]
-/// ```
-/// use objc2::encode::Encode;
-/// <()>::ENCODING; // TODO: Make this fail as well
-/// ```
-/// ```should_fail
-/// use core::ffi::c_void;
-/// use objc2::encode::Encode;
-/// <c_void>::ENCODING;
-/// ```
-/// ```should_fail
-/// use objc2::encode::Encode;
-/// <*const ()>::ENCODING;
-/// ```
-/// ```should_fail
-/// use core::ffi::c_void;
-/// use objc2::encode::Encode;
-/// <&c_void>::ENCODING;
-/// ```
-extern "C" {}
 
 macro_rules! encode_impls_size {
     ($($t:ty => ($t16:ty, $t32:ty, $t64:ty),)*) => ($(
@@ -606,22 +576,22 @@ unsafe impl<T: RefEncode + ?Sized> OptionEncode for NonNull<T> {}
 /// pointers that are higher-ranked over lifetimes don't get implemented.
 ///
 /// We could fix it by adding those impls and allowing `coherence_leak_check`,
-/// but it would have to be done for _all_ references, `Option<&T>` and such as
-/// well. So trying to do it quickly requires generating a polynomial amount of
-/// implementations, which IMO is overkill for such a small issue.
+/// but it would have to be done for _all_ references, `Option<&T>` and such
+/// as well. So trying to do it quickly requires generating a polynomial
+/// amount of implementations, which IMO is overkill for such a small issue.
 ///
 /// Using `?Sized` is probably not safe here because C functions can only take
 /// and return items with a known size.
 macro_rules! encode_fn_pointer_impl {
     (@ $FnTy: ty, $($Arg: ident),*) => {
-        unsafe impl<Ret: Encode, $($Arg: Encode),*> Encode for $FnTy {
+        unsafe impl<Ret: __unstable::EncodeReturn, $($Arg: Encode),*> Encode for $FnTy {
             const ENCODING: Encoding = Encoding::Pointer(&Encoding::Unknown);
         }
-        unsafe impl<Ret: Encode, $($Arg: Encode),*> RefEncode for $FnTy {
+        unsafe impl<Ret: __unstable::EncodeReturn, $($Arg: Encode),*> RefEncode for $FnTy {
             const ENCODING_REF: Encoding = Encoding::Pointer(&Self::ENCODING);
         }
         // SAFETY: Function pointers have a NULL niche
-        unsafe impl<Ret: Encode, $($Arg: Encode),*> OptionEncode for $FnTy {}
+        unsafe impl<Ret: __unstable::EncodeReturn, $($Arg: Encode),*> OptionEncode for $FnTy {}
     };
     (# $abi:literal; $($Arg: ident),+) => {
         // Normal functions
@@ -657,134 +627,10 @@ encode_fn_pointer_impl!(A, B, C, D, E, F, G, H, I, J);
 encode_fn_pointer_impl!(A, B, C, D, E, F, G, H, I, J, K);
 encode_fn_pointer_impl!(A, B, C, D, E, F, G, H, I, J, K, L);
 
-mod convert_private {
-    use super::*;
-
-    pub trait Sealed {}
-    impl<T: Encode> Sealed for T {}
-    impl Sealed for bool {}
-}
-
-/// Represents types that can easily be converted to/from an [`Encode`] type.
-///
-/// This is implemented specially for [`bool`] to allow using that as
-/// Objective-C `BOOL`, where it would otherwise not be allowed (since they
-/// are not ABI compatible).
-///
-/// This is mostly an implementation detail, and hence the trait is sealed.
-/// Open an issue if you know a use-case where this restrition should be
-/// lifted!
-pub trait EncodeConvert: convert_private::Sealed {
-    /// The inner type that this can be converted to and from.
-    #[doc(hidden)]
-    type __Inner: Encode;
-
-    /// The actual encoding this type has.
-    #[doc(hidden)]
-    const __ENCODING: Encoding;
-
-    #[doc(hidden)]
-    fn __from_inner(inner: Self::__Inner) -> Self;
-
-    #[doc(hidden)]
-    fn __into_inner(self) -> Self::__Inner;
-}
-
-impl<T: Encode> EncodeConvert for T {
-    type __Inner = Self;
-    const __ENCODING: Encoding = Self::ENCODING;
-
-    #[inline]
-    fn __from_inner(inner: Self::__Inner) -> Self {
-        inner
-    }
-
-    #[inline]
-    fn __into_inner(self) -> Self::__Inner {
-        self
-    }
-}
-
-impl EncodeConvert for bool {
-    type __Inner = Bool;
-    const __ENCODING: Encoding = Encoding::Bool;
-
-    #[inline]
-    fn __from_inner(inner: Self::__Inner) -> Self {
-        inner.as_bool()
-    }
-
-    #[inline]
-    fn __into_inner(self) -> Self::__Inner {
-        Bool::new(self)
-    }
-}
-
-mod args_private {
-    pub trait Sealed {}
-}
-
-/// Types that represent an ordered group of function arguments, where each
-/// argument has an Objective-C type-encoding, or can be converted from one.
-///
-/// This is implemented for tuples of up to 16 arguments, where each argument
-/// implements [`EncodeConvert`]. It is primarily used to make generic code
-/// a bit easier.
-///
-/// Note that tuples themselves don't implement [`Encode`] directly, because
-/// they're not FFI-safe!
-///
-/// This is a sealed trait, and should not need to be implemented. Open an
-/// issue if you know a use-case where this restrition should be lifted!
-///
-///
-/// # Safety
-///
-/// You cannot rely on this trait for ensuring that the arguments all
-/// are [`Encode`], since they may only implement [`EncodeConvert`]. Use your
-/// own trait and add [`Encode`] bounds on that.
-pub unsafe trait EncodeArguments: args_private::Sealed {
-    /// The encodings for the arguments.
-    const ENCODINGS: &'static [Encoding];
-}
-
-macro_rules! encode_args_impl {
-    ($($Arg: ident),*) => {
-        impl<$($Arg: EncodeConvert),*> args_private::Sealed for ($($Arg,)*) {}
-
-        unsafe impl<$($Arg: EncodeConvert),*> EncodeArguments for ($($Arg,)*) {
-            const ENCODINGS: &'static [Encoding] = &[
-                // T::__Inner::ENCODING => T::ENCODING
-                // bool::__Inner::ENCODING => Bool::ENCODING
-                $($Arg::__Inner::ENCODING),*
-            ];
-        }
-    };
-}
-
-encode_args_impl!();
-encode_args_impl!(A);
-encode_args_impl!(A, B);
-encode_args_impl!(A, B, C);
-encode_args_impl!(A, B, C, D);
-encode_args_impl!(A, B, C, D, E);
-encode_args_impl!(A, B, C, D, E, F);
-encode_args_impl!(A, B, C, D, E, F, G);
-encode_args_impl!(A, B, C, D, E, F, G, H);
-encode_args_impl!(A, B, C, D, E, F, G, H, I);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J, K);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J, K, L);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J, K, L, M);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J, K, L, M, N);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J, K, L, M, N, O);
-encode_args_impl!(A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P);
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
-    use core::any::TypeId;
     use core::sync::atomic::*;
 
     #[test]
@@ -839,8 +685,6 @@ mod tests {
 
     #[test]
     fn test_void() {
-        // TODO: Remove this
-        assert_eq!(<()>::ENCODING, Encoding::Void);
         assert_eq!(
             <*const c_void>::ENCODING,
             Encoding::Pointer(&Encoding::Void)
@@ -874,7 +718,7 @@ mod tests {
             Encoding::Pointer(&Encoding::Unknown)
         );
         assert_eq!(
-            <extern "C" fn(x: ()) -> ()>::ENCODING,
+            <extern "C" fn(x: i32) -> u32>::ENCODING,
             Encoding::Pointer(&Encoding::Unknown)
         );
         assert_eq!(
@@ -905,50 +749,5 @@ mod tests {
         impls_encode(my_fn2 as extern "C" fn(_, _));
         impls_encode(my_fn3 as extern "C" fn(_) -> _);
         impls_encode(my_fn4 as extern "C" fn(_, _) -> _);
-    }
-
-    #[test]
-    fn convert_normally_noop() {
-        assert_eq!(
-            TypeId::of::<<i32 as EncodeConvert>::__Inner>(),
-            TypeId::of::<i32>()
-        );
-        assert_eq!(<i32 as EncodeConvert>::__from_inner(42), 42);
-        assert_eq!(42i32.__into_inner(), 42);
-    }
-
-    #[test]
-    fn convert_i8() {
-        assert_eq!(
-            TypeId::of::<<i8 as EncodeConvert>::__Inner>(),
-            TypeId::of::<i8>()
-        );
-        assert_eq!(<i8 as EncodeConvert>::__from_inner(-3), -3);
-        assert_eq!((-3i32).__into_inner(), -3);
-    }
-
-    #[test]
-    fn convert_bool() {
-        assert!(!<bool as EncodeConvert>::__from_inner(Bool::NO));
-        assert!(<bool as EncodeConvert>::__from_inner(Bool::YES));
-
-        assert!(!false.__into_inner().as_bool());
-        assert!(true.__into_inner().as_bool());
-        assert_eq!(bool::__ENCODING, Encoding::Bool);
-
-        assert_eq!(
-            <bool as EncodeConvert>::__Inner::__ENCODING,
-            <bool as EncodeConvert>::__Inner::ENCODING
-        );
-
-        #[cfg(all(feature = "apple", target_os = "macos", target_arch = "x86_64"))]
-        assert_eq!(<bool as EncodeConvert>::__Inner::ENCODING, Encoding::Char);
-    }
-
-    #[test]
-    fn test_encode_arguments() {
-        assert!(<()>::ENCODINGS.is_empty());
-        assert_eq!(<(i8,)>::ENCODINGS, &[i8::ENCODING]);
-        assert_eq!(<(i8, u32)>::ENCODINGS, &[i8::ENCODING, u32::ENCODING]);
     }
 }