+++ crates/objc2/src/macros/declare_class.rs
@@ -84,6 +84,9 @@
 /// make it behave similarly to the Objective-C `BOOL`. Use [`runtime::Bool`]
 /// if you want to control this manually.
 ///
+/// Note that `&mut Id<_, _>` and other such out parameters are not yet
+/// supported, and may generate a panic at runtime.
+///
 /// ["associated functions"]: https://doc.rust-lang.org/reference/items/associated-items.html#methods
 /// ["methods"]: https://doc.rust-lang.org/reference/items/associated-items.html#methods
 /// [`extern_methods!`]: crate::extern_methods
@@ -879,7 +882,7 @@ macro_rules! __declare_class_rewrite_args {
     } => {
         $crate::__declare_class_rewrite_args! {
             ($($($rest_args)*)?)
-            ($($args_converted)* _ : <$param_ty as $crate::encode::EncodeConvert>::__Inner,)
+            ($($args_converted)* _ : <$param_ty as $crate::encode::__unstable::EncodeConvertArgument>::__Inner,)
             ($($body_prefix)*)
 
             ($out_macro)
@@ -897,10 +900,10 @@ macro_rules! __declare_class_rewrite_args {
     } => {
         $crate::__declare_class_rewrite_args! {
             ($($($rest_args)*)?)
-            ($($args_converted)* $param : <$param_ty as $crate::encode::EncodeConvert>::__Inner,)
+            ($($args_converted)* $param : <$param_ty as $crate::encode::__unstable::EncodeConvertArgument>::__Inner,)
             (
                 $($body_prefix)*
-                let mut $param = <$param_ty as $crate::encode::EncodeConvert>::__from_inner($param);
+                let mut $param = <$param_ty as $crate::encode::__unstable::EncodeConvertArgument>::__from_declared_param($param);
             )
 
             ($out_macro)
@@ -918,10 +921,10 @@ macro_rules! __declare_class_rewrite_args {
     } => {
         $crate::__declare_class_rewrite_args! {
             ($($($rest_args)*)?)
-            ($($args_converted)* $param : <$param_ty as $crate::encode::EncodeConvert>::__Inner,)
+            ($($args_converted)* $param : <$param_ty as $crate::encode::__unstable::EncodeConvertArgument>::__Inner,)
             (
                 $($body_prefix)*
-                let $param = <$param_ty as $crate::encode::EncodeConvert>::__from_inner($param);
+                let $param = <$param_ty as $crate::encode::__unstable::EncodeConvertArgument>::__from_declared_param($param);
             )
 
             ($out_macro)
@@ -972,7 +975,7 @@ macro_rules! __declare_class_method_out_inner {
         $($qualifiers)* extern "C" fn $name(
             $($args_prefix)*
             $($args_converted)*
-        ) $(-> <$ret as $crate::encode::EncodeConvert>::__Inner)? {
+        ) $(-> <$ret as $crate::encode::__unstable::EncodeConvertReturn>::__Inner)? {
             $($body_prefix)*
             $crate::__convert_result! {
                 $body $(; $ret)?
@@ -1057,7 +1060,9 @@ macro_rules! __convert_result {
     ($body:block; $ret:ty) => {
         let __objc2_result = $body;
         #[allow(unreachable_code)]
-        <$ret as $crate::encode::EncodeConvert>::__into_inner(__objc2_result)
+        <$ret as $crate::encode::__unstable::EncodeConvertReturn>::__into_declared_return(
+            __objc2_result,
+        )
     };
 }